Move to PyQt6
This breaks PyQt5/PyQt4, but those could perhaps be fixed, albeit
the enum changes will be painful.

Major change is fixed to use appropriate enums which have been introduced.

And redo import/references of QtGui from earlier split into QtWidgets
that came with PyQt5. This makes a Widgets class object that could
be used to still support older PyQt.  Or perhaps should have left this
part as is.

References that helped in fixing enums:
https://docs.w3cub.com/qt~5.15/
https://doc.qt.io/qt-5/qtgui-module.html
https://doc.qt.io/qtforpython/PySide6/QtCore/
https://doc.qt.io/qtforpython/PySide6/QtWidgets/